web_backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

mysqlclient==2.2.8
PyMySQL==1.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.8.0
surprise==0.1
joblib==1.3.2
jieba==0.42.1